sfu update readme.md
hi there,
i'm taking a CS writing course in SFU. my job is to review other documents on github. the description of you application has a redundant issue. its a Go application, so it is unnecessary to say "written in GO". For example, it's a java app, written in java. therefore, i rewrote it.